+++ core/src/main/java/org/everit/json/schema/event/CombinedSchemaMatchEvent.java
@@ -1,4 +1,4 @@
-package org.everit.json.schema.listener;
+package org.everit.json.schema.event;
 
 import org.everit.json.schema.CombinedSchema;
 import org.everit.json.schema.Schema;
