+++ core/src/main/java/org/everit/json/schema/event/MismatchEvent.java
@@ -1,4 +1,4 @@
-package org.everit.json.schema.listener;
+package org.everit.json.schema.event;
 
 import org.everit.json.schema.ValidationException;
