+++ core/src/test/java/org/everit/json/schema/ConditionalSchemaEventsTest.java
@@ -3,15 +3,15 @@ package org.everit.json.schema;
 import static org.everit.json.schema.ConditionalSchemaTest.MAX_LENGTH_STRING_SCHEMA;
 import static org.everit.json.schema.ConditionalSchemaTest.MIN_LENGTH_STRING_SCHEMA;
 import static org.everit.json.schema.ConditionalSchemaTest.PATTERN_STRING_SCHEMA;
-import static org.everit.json.schema.listener.ConditionalSchemaValidationEvent.Keyword.ELSE;
-import static org.everit.json.schema.listener.ConditionalSchemaValidationEvent.Keyword.IF;
-import static org.everit.json.schema.listener.ConditionalSchemaValidationEvent.Keyword.THEN;
+import static org.everit.json.schema.event.ConditionalSchemaValidationEvent.Keyword.ELSE;
+import static org.everit.json.schema.event.ConditionalSchemaValidationEvent.Keyword.IF;
+import static org.everit.json.schema.event.ConditionalSchemaValidationEvent.Keyword.THEN;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 
-import org.everit.json.schema.listener.ConditionalSchemaMatchEvent;
-import org.everit.json.schema.listener.ConditionalSchemaMismatchEvent;
-import org.everit.json.schema.listener.ValidationListener;
+import org.everit.json.schema.event.ConditionalSchemaMatchEvent;
+import org.everit.json.schema.event.ConditionalSchemaMismatchEvent;
+import org.everit.json.schema.event.ValidationListener;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
