+++ core/src/test/java/org/everit/json/schema/ValidatingVisitorTest.java
@@ -10,9 +10,9 @@ import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 
-import org.everit.json.schema.listener.CombinedSchemaMatchEvent;
-import org.everit.json.schema.listener.CombinedSchemaMismatchEvent;
-import org.everit.json.schema.listener.ValidationListener;
+import org.everit.json.schema.event.CombinedSchemaMatchEvent;
+import org.everit.json.schema.event.CombinedSchemaMismatchEvent;
+import org.everit.json.schema.event.ValidationListener;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.Before;
