+++ core/src/test/resources/org/everit/jsonvalidator/event/combined-schema-match.json
