+++ core/src/test/resources/org/everit/jsonvalidator/event/conditional-match-if.json
