+++ core/src/test/resources/org/everit/jsonvalidator/event/conditional-match-then-boolean.json
