+++ core/src/test/resources/org/everit/jsonvalidator/event/conditional-mismatch-if.json
