+++ core/src/test/resources/org/everit/jsonvalidator/event/reference-event-with-instance.json
